# Optimize DataFetchingSelectionSet.getImmediateFields() to avoid traversing descendants