--- a/src/main/java/graphql/schema/DataFetchingFieldSelectionSetImpl.java
+++ b/src/main/java/graphql/schema/DataFetchingFieldSelectionSetImpl.java
@@ -90,7 +90,10 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
 
     private final Supplier<ExecutableNormalizedField> normalizedFieldSupplier;
 
-    private LockKit.ComputedOnce computedOnce = new LockKit.ComputedOnce();
+    private final LockKit.ReentrantLock lock = new LockKit.ReentrantLock();
+    private volatile boolean computedValues;
+    private volatile boolean computedImmediateValues;
+
     // we have multiple entries in this map so that we can do glob matching in multiple ways
     // however it needs to be normalised back to a set of unique fields when give back out to
     // the caller.
@@ -109,7 +112,7 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
         if (fieldGlobPattern == null || fieldGlobPattern.isEmpty()) {
             return false;
         }
-        computeValuesLazily();
+        computeValuesLazily(false);
         fieldGlobPattern = removeLeadingSlash(fieldGlobPattern);
         PathMatcher globMatcher = globMatcher(fieldGlobPattern);
         for (String flattenedField : flattenedFieldsForGlobSearching) {
@@ -159,7 +162,7 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
         if (fieldGlobPattern == null || fieldGlobPattern.isEmpty()) {
             return emptyList();
         }
-        computeValuesLazily();
+        computeValuesLazily(false);
 
         List<String> targetNames = new ArrayList<>();
         for (String flattenedField : flattenedFieldsForGlobSearching) {
@@ -178,7 +181,7 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
 
     @Override
     public List<SelectedField> getFields() {
-        computeValuesLazily();
+        computeValuesLazily(false);
         return toSetSemanticsList(normalisedSelectionSetFields.values().stream()
                 .flatMap(Collection::stream));
     }
@@ -190,7 +193,7 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
 
     @Override
     public List<SelectedField> getImmediateFields() {
-        computeValuesLazily();
+        computeValuesLazily(true);
         return immediateFields;
     }
 
@@ -204,24 +207,40 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
         return getFields(fieldGlobPattern, fieldGlobPatterns).stream().collect(Collectors.groupingBy(SelectedField::getResultKey));
     }
 
-    private void computeValuesLazily() {
-        if (computedOnce.hasBeenComputed()) {
+    private void computeValuesLazily(boolean immediate) {
+        if (computedValues) {
             return;
         }
+
+        // Avoid recomputing the immediate fields if they have already been computed.
+        if (immediate && computedImmediateValues) {
+            return;
+        }
+
         // this supplier is a once only thread synced call - so do it outside this lock
         // if only to have only 1 lock in action at a time
         ExecutableNormalizedField currentNormalisedField = normalizedFieldSupplier.get();
-        computedOnce.runOnce(() -> {
+
+        lock.runLocked(() -> {
+            if (computedValues) {
+                return;
+            }
+
+            if (computedImmediateValues && immediate) {
+                return;
+            }
+
             flattenedFieldsForGlobSearching = new LinkedHashSet<>();
             normalisedSelectionSetFields = new LinkedHashMap<>();
             ImmutableList.Builder<SelectedField> immediateFieldsBuilder = ImmutableList.builder();
-            traverseSubSelectedFields(currentNormalisedField, immediateFieldsBuilder, "", "", true);
+            traverseSubSelectedFields(currentNormalisedField, immediateFieldsBuilder, "", "", true, immediate);
             immediateFields = immediateFieldsBuilder.build();
+            computedImmediateValues = true;
+            computedValues = !immediate;
         });
     }
 
-
-    private void traverseSubSelectedFields(ExecutableNormalizedField currentNormalisedField, ImmutableList.Builder<SelectedField> immediateFieldsBuilder, String qualifiedFieldPrefix, String simpleFieldPrefix, boolean firstLevel) {
+    private void traverseSubSelectedFields(ExecutableNormalizedField currentNormalisedField, ImmutableList.Builder<SelectedField> immediateFieldsBuilder, String qualifiedFieldPrefix, String simpleFieldPrefix, boolean firstLevel, boolean immediate) {
         List<ExecutableNormalizedField> children = currentNormalisedField.getChildren();
         for (ExecutableNormalizedField normalizedSubSelectedField : children) {
             String typeQualifiedName = mkTypeQualifiedName(normalizedSubSelectedField);
@@ -241,8 +260,8 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
             normalisedSelectionSetFields.computeIfAbsent(globQualifiedName, newList()).add(selectedField);
             normalisedSelectionSetFields.computeIfAbsent(globSimpleName, newList()).add(selectedField);
 
-            if (normalizedSubSelectedField.hasChildren()) {
-                traverseSubSelectedFields(normalizedSubSelectedField, immediateFieldsBuilder, globQualifiedName, globSimpleName, false);
+            if (normalizedSubSelectedField.hasChildren() && !immediate) {
+                traverseSubSelectedFields(normalizedSubSelectedField, immediateFieldsBuilder, globQualifiedName, globSimpleName, false, false);
             }
         }
     }
@@ -276,7 +295,7 @@ public class DataFetchingFieldSelectionSetImpl implements DataFetchingFieldSelec
 
     @Override
     public String toString() {
-        if (!computedOnce.hasBeenComputed()) {
+        if (!computedValues) {
             return "notComputed";
         }
         return String.join("\n", flattenedFieldsForGlobSearching);