Addresses #3854.

## Description

Computes immediate fields only for `getImmediateFields()` to avoid cost of traversing the entire sub-tree to build glob-patterns for descendant fields.

## Benchmark
Before:
```
Benchmark                                                          Mode  Cnt   Score   Error   Units
DFSelectionSetPerformance.benchMarkThroughput                     thrpt    9  10.424 ± 0.598  ops/ms
DFSelectionSetPerformance.benchMarkThroughput_getImmediateFields  thrpt    9  21.992 ± 0.224  ops/ms
DFSelectionSetPerformance.benchMarkAvgTime                         avgt    9   0.096 ± 0.005   ms/op
DFSelectionSetPerformance.benchMarkAvgTime_getImmediateFields      avgt    9   0.044 ± 0.001   ms/op
```

After:
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
DFSelectionSetPerformance.benchMarkThroughput                     thrpt    9    10.641 ±  0.118  ops/ms
DFSelectionSetPerformance.benchMarkThroughput_getImmediateFields  thrpt    9  2667.783 ± 73.924  ops/ms
DFSelectionSetPerformance.benchMarkAvgTime                         avgt    9     0.094 ±  0.003   ms/op
DFSelectionSetPerformance.benchMarkAvgTime_getImmediateFields      avgt    9    ≈ 10⁻³            ms/op
```