@timward60 2025-03-18
> Hey @timward60 ... thanks a lot for the PR and even more for the Benchmark.
> 
> I have a request:
> 
> We just introduced automated performance tests, which enables us to track the performance much better over time as it is executed in a stable reproducible environment.
> 
> Can you create another PR which moves the `DFSelectionSetBenchmark` into the `performance` package and renames into `DFSelectionSetPerformance`?
> 
> We can merge this PR first and it gives us a clear baseline for the improvements. Thanks a lot

Nice to see, created the PR here: #3856