@andimarek 2025-03-19
@timward60 you can see the performance results from master tests here: https://github.com/graphql-java/graphql-java/tree/master/performance-results